fix: moved out coursesUnlockedWhenTaken, and consolidated courses.ts into course.ts
